pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "marklogic-mcp"
version = "0.1.0.dev3"
description = "A Python-based MCP server for MarkLogic database operations"
readme = "README.md"
authors = [
    {name = "Karthik", email = "karthiknarayan@graphai.ca"}  # Please provide your actual email
]
requires-python = ">=3.7"
keywords = ["marklogic", "database", "mcp", "server"]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "marklogic-python-client>=1.0.0",
    "fastmcp>=0.1.0",
    "toml>=0.10.2",
    "requests>=2.31.0"
]

[project.urls]
Homepage = "https://github.com/karthiknarayankotha/marklogic-mcp"
Repository = "https://github.com/karthiknarayankotha/marklogic-mcp.git"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.marklogic]
host = "localhost"
port = 8000
user = "admin"
password = "admin"

[tool.marklogic.logging]
level = "INFO"
format = "%(asctime)s - %(name)s - %(levelname)s - %(message)s" 